Move tests utils to a separate file
